Pass result of realize() thru DelayedArray() constructor
This is to make sure to always produce a DelayedArray object or derivative,
even after the following planned change for the default realize() method.

Planned change for the default realize() method: The plan is to change
what 'realize( , BACKEND=NULL)' does. Instead of returning an ordinary array
wrapped in a DelayedArray object like it does at the moment, it will return
a **naked** ordinary array. This change is planned for DelayedArray 0.29.1.